+++ src/checks/incorrect_delimiter.rs
@@ -27,13 +27,14 @@ impl Check for IncorrectDelimiterChecker {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use std::path::PathBuf;
 
     #[test]
     fn working_run() {
         let mut checker = IncorrectDelimiterChecker::default();
         let line = LineEntry {
             number: 1,
-            file_name: String::from(".env"),
+            file_path: PathBuf::from(".env"),
             raw_string: String::from("FOO_BAR=FOOBAR"),
         };
         assert_eq!(None, checker.run(line));
@@ -44,7 +45,7 @@ mod tests {
         let mut checker = IncorrectDelimiterChecker::default();
         let line = LineEntry {
             number: 1,
-            file_name: String::from(".env"),
+            file_path: PathBuf::from(".env"),
             raw_string: String::from("F1OO=BAR"),
         };
         assert_eq!(None, checker.run(line));
@@ -55,7 +56,7 @@ mod tests {
         let mut checker = IncorrectDelimiterChecker::default();
         let line = LineEntry {
             number: 1,
-            file_name: String::from(".env"),
+            file_path: PathBuf::from(".env"),
             raw_string: String::from("FOO-BAR=FOOBAR"),
         };
         let expected = Some(Warning::new(
@@ -70,7 +71,7 @@ mod tests {
         let mut checker = IncorrectDelimiterChecker::default();
         let line = LineEntry {
             number: 1,
-            file_name: String::from(".env"),
+            file_path: PathBuf::from(".env"),
             raw_string: String::from("FOO BAR=FOOBAR"),
         };
         let expected = Some(Warning::new(
@@ -85,7 +86,7 @@ mod tests {
         let mut checker = IncorrectDelimiterChecker::default();
         let line = LineEntry {
             number: 1,
-            file_name: String::from(".env"),
+            file_path: PathBuf::from(".env"),
             raw_string: String::from("FOO-BAR"),
         };
         assert_eq!(None, checker.run(line));
@@ -96,7 +97,7 @@ mod tests {
         let mut checker = IncorrectDelimiterChecker::default();
         let line = LineEntry {
             number: 1,
-            file_name: String::from(".env"),
+            file_path: PathBuf::from(".env"),
             raw_string: String::from(" FOO=FOOBAR"),
         };
         assert_eq!(None, checker.run(line));
@@ -107,7 +108,7 @@ mod tests {
         let mut checker = IncorrectDelimiterChecker::default();
         let line = LineEntry {
             number: 1,
-            file_name: String::from(".env"),
+            file_path: PathBuf::from(".env"),
             raw_string: String::from("FOO_BAR =FOOBAR"),
         };
         assert_eq!(None, checker.run(line));
@@ -118,7 +119,7 @@ mod tests {
         let mut checker = IncorrectDelimiterChecker::default();
         let line = LineEntry {
             number: 1,
-            file_name: String::from(".env"),
+            file_path: PathBuf::from(".env"),
             raw_string: String::from(""),
         };
         assert_eq!(None, checker.run(line));
@@ -129,7 +130,7 @@ mod tests {
         let mut checker = IncorrectDelimiterChecker::default();
         let line = LineEntry {
             number: 1,
-            file_name: String::from(".env"),
+            file_path: PathBuf::from(".env"),
             raw_string: String::from("F=BAR"),
         };
         assert_eq!(None, checker.run(line));