+++ src/checks/leading_character.rs
@@ -29,6 +29,7 @@ impl Check for LeadingCharacterChecker {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use std::path::PathBuf;
 
     const MESSAGE: &str = "Invalid leading character detected";
 
@@ -37,7 +38,7 @@ mod tests {
         let mut checker = LeadingCharacterChecker::default();
         let line = LineEntry {
             number: 1,
-            file_name: String::from(".env"),
+            file_path: PathBuf::from(".env"),
             raw_string: String::from("FOO=BAR"),
         };
         assert_eq!(None, checker.run(line));
@@ -48,7 +49,7 @@ mod tests {
         let mut checker = LeadingCharacterChecker::default();
         let line = LineEntry {
             number: 1,
-            file_name: String::from(".env"),
+            file_path: PathBuf::from(".env"),
             raw_string: String::from("_FOO=BAR"),
         };
         assert_eq!(None, checker.run(line));
@@ -59,7 +60,7 @@ mod tests {
         let mut checker = LeadingCharacterChecker::default();
         let line = LineEntry {
             number: 1,
-            file_name: String::from(".env"),
+            file_path: PathBuf::from(".env"),
             raw_string: String::from(".FOO=BAR"),
         };
         assert_eq!(
@@ -73,7 +74,7 @@ mod tests {
         let mut checker = LeadingCharacterChecker::default();
         let line = LineEntry {
             number: 1,
-            file_name: String::from(".env"),
+            file_path: PathBuf::from(".env"),
             raw_string: String::from("*FOO=BAR"),
         };
         assert_eq!(
@@ -87,7 +88,7 @@ mod tests {
         let mut checker = LeadingCharacterChecker::default();
         let line = LineEntry {
             number: 1,
-            file_name: String::from(".env"),
+            file_path: PathBuf::from(".env"),
             raw_string: String::from("1FOO=BAR"),
         };
         assert_eq!(
@@ -101,7 +102,7 @@ mod tests {
         let mut checker = LeadingCharacterChecker::default();
         let line = LineEntry {
             number: 1,
-            file_name: String::from(".env"),
+            file_path: PathBuf::from(".env"),
             raw_string: String::from(" FOO=BAR"),
         };
         let expected = Some(Warning::new(line.clone(), MESSAGE.to_string()));
@@ -113,7 +114,7 @@ mod tests {
         let mut checker = LeadingCharacterChecker::default();
         let line = LineEntry {
             number: 1,
-            file_name: String::from(".env"),
+            file_path: PathBuf::from(".env"),
             raw_string: String::from("  FOO=BAR"),
         };
         let expected = Some(Warning::new(line.clone(), MESSAGE.to_string()));
@@ -125,7 +126,7 @@ mod tests {
         let mut checker = LeadingCharacterChecker::default();
         let line = LineEntry {
             number: 1,
-            file_name: String::from(".env"),
+            file_path: PathBuf::from(".env"),
             raw_string: String::from("\tFOO=BAR"),
         };
         let expected = Some(Warning::new(line.clone(), MESSAGE.to_string()));