+++ src/checks/spaces_around_equal.rs
@@ -30,6 +30,7 @@ impl Check for SpacesAroundEqualChecker {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use std::path::PathBuf;
 
     const MESSAGE: &str = "The line has spaces around equal sign";
 
@@ -38,7 +39,7 @@ mod tests {
         let mut checker = SpacesAroundEqualChecker::default();
         let line = LineEntry {
             number: 1,
-            file_name: String::from(".env"),
+            file_path: PathBuf::from(".env"),
             raw_string: String::from("DEBUG_HTTP=true"),
         };
         assert_eq!(None, checker.run(line));
@@ -49,7 +50,7 @@ mod tests {
         let mut checker = SpacesAroundEqualChecker::default();
         let line = LineEntry {
             number: 1,
-            file_name: String::from(".env"),
+            file_path: PathBuf::from(".env"),
             raw_string: String::from(" DEBUG_HTTP=true"),
         };
         assert_eq!(None, checker.run(line));
@@ -60,7 +61,7 @@ mod tests {
         let mut checker = SpacesAroundEqualChecker::default();
         let line = LineEntry {
             number: 1,
-            file_name: String::from(".env"),
+            file_path: PathBuf::from(".env"),
             raw_string: String::from("DEBUG_HTTP=true "),
         };
         assert_eq!(None, checker.run(line));
@@ -71,7 +72,7 @@ mod tests {
         let mut checker = SpacesAroundEqualChecker::default();
         let line = LineEntry {
             number: 1,
-            file_name: String::from(".env"),
+            file_path: PathBuf::from(".env"),
             raw_string: String::from(""),
         };
         assert_eq!(None, checker.run(line));
@@ -82,7 +83,7 @@ mod tests {
         let mut checker = SpacesAroundEqualChecker::default();
         let line = LineEntry {
             number: 1,
-            file_name: String::from(".env"),
+            file_path: PathBuf::from(".env"),
             raw_string: String::from("DEBUG_HTTP true"),
         };
         assert_eq!(None, checker.run(line));
@@ -93,7 +94,7 @@ mod tests {
         let mut checker = SpacesAroundEqualChecker::default();
         let line = LineEntry {
             number: 1,
-            file_name: String::from(".env"),
+            file_path: PathBuf::from(".env"),
             raw_string: String::from("DEBUG-HTTP = true"),
         };
         let expected = Some(Warning::new(line.clone(), MESSAGE.to_string()));
@@ -105,7 +106,7 @@ mod tests {
         let mut checker = SpacesAroundEqualChecker::default();
         let line = LineEntry {
             number: 1,
-            file_name: String::from(".env"),
+            file_path: PathBuf::from(".env"),
             raw_string: String::from("DEBUG-HTTP =true"),
         };
         let expected = Some(Warning::new(line.clone(), MESSAGE.to_string()));
@@ -117,7 +118,7 @@ mod tests {
         let mut checker = SpacesAroundEqualChecker::default();
         let line = LineEntry {
             number: 1,
-            file_name: String::from(".env"),
+            file_path: PathBuf::from(".env"),
             raw_string: String::from("DEBUG-HTTP= true"),
         };
         let expected = Some(Warning::new(line.clone(), MESSAGE.to_string()));