+++ src/lib.rs
@@ -1,4 +1,5 @@
 use crate::common::*;
+
 use clap::Arg;
 use std::env;
 use std::error::Error;
@@ -10,12 +11,20 @@ use std::path::PathBuf;
 mod checks;
 mod common;
 
-fn get_args<'a>() -> clap::ArgMatches<'a> {
+fn get_args(current_dir: &str) -> clap::ArgMatches {
     clap::App::new(env!("CARGO_PKG_NAME"))
         .about(env!("CARGO_PKG_DESCRIPTION"))
         .author(env!("CARGO_PKG_AUTHORS"))
         .version(env!("CARGO_PKG_VERSION"))
         .version_short("v")
+        .arg(
+            Arg::with_name("input")
+                .help("files or paths")
+                .index(1)
+                .default_value(current_dir)
+                .required(true)
+                .multiple(true),
+        )
         .arg(
             Arg::with_name("include")
                 .short("i")
@@ -46,18 +55,75 @@ fn get_args<'a>() -> clap::ArgMatches<'a> {
         .get_matches()
 }
 
+#[allow(clippy::redundant_closure)]
 pub fn run() -> Result<Vec<Warning>, Box<dyn Error>> {
-    let args = get_args();
-
-    let dir_path = match args.value_of("path") {
-        Some(path) => PathBuf::from(path),
-        None => env::current_dir()?,
+    let current_dir = match env::current_dir() {
+        Ok(dir) => dir,
+        Err(e) => return Err(Box::new(e)),
     };
 
-    let files = dotenv_files(args, dir_path)?;
-    let mut warnings: Vec<Warning> = Vec::new();
+    let current_dir = current_dir.to_str().unwrap();
 
+    let args = get_args(current_dir);
+
+    let mut paths: Vec<PathBuf> = vec![];
+    let mut files: Vec<FileEntry> = vec![];
+
+    if let Some(inputs) = args.values_of("input") {
+        paths = inputs
+            .clone()
+            .filter_map(|s| fs::canonicalize(s).ok())
+            .filter(|p| p.is_dir())
+            .collect();
+
+        files = inputs
+            .filter_map(|s| fs::canonicalize(s).ok())
+            .filter(|p| p.is_file())
+            .filter_map(|p| FileEntry::from(p))
+            .collect();
+    }
+
+    for path in paths {
+        let entries = path.read_dir()?;
+
+        let mut file_paths: Vec<FileEntry> = entries
+            .filter_map(|e| e.ok())
+            .filter_map(|e| FileEntry::from(e.path()))
+            .filter(|f| f.is_env_file())
+            .collect();
+
+        files.append(file_paths.as_mut());
+    }
+
+    // Removes files from paths if they should be excluded
+    if let Some(excluded) = args.values_of("exclude") {
+        let excluded_paths: Vec<PathBuf> =
+            excluded.filter_map(|f| fs::canonicalize(f).ok()).collect();
+
+        files.retain(|f| !excluded_paths.contains(&f.path));
+    }
+
+    files.sort();
+    files.dedup();
+
+    let mut new_files: Vec<FileEntry> = vec![];
     for file in files {
+        let result = file.path.strip_prefix(&current_dir);
+        let strip_path = match result {
+            Ok(p) => p,
+            Err(_) => continue,
+        };
+
+        let new_file = match FileEntry::from(strip_path.to_owned()) {
+            Some(f) => f,
+            None => continue,
+        };
+
+        new_files.push(new_file);
+    }
+
+    let mut warnings: Vec<Warning> = Vec::new();
+    for file in new_files {
         let f = File::open(&file.path)?;
         let reader = BufReader::new(f);
 
@@ -68,8 +134,8 @@ pub fn run() -> Result<Vec<Warning>, Box<dyn Error>> {
             let raw_string = line?;
 
             lines.push(LineEntry {
-                file_name: file.file_name.clone(),
                 number,
+                file_path: file.path.clone(),
                 raw_string,
             })
         }
@@ -79,41 +145,4 @@ pub fn run() -> Result<Vec<Warning>, Box<dyn Error>> {
     }
 
     Ok(warnings)
-}
-
-/// Returns `Result<Vec<FileEntry>` of files with the the `.env` prefix
-#[allow(clippy::redundant_closure)]
-fn dotenv_files(
-    args: clap::ArgMatches,
-    dir_path: PathBuf,
-) -> Result<Vec<FileEntry>, Box<dyn Error>> {
-    let entries = dir_path.read_dir()?;
-
-    let mut paths: Vec<FileEntry> = entries
-        .filter_map(|e| e.ok())
-        .filter_map(|f| fs::canonicalize(f.path()).ok())
-        .filter_map(|p| FileEntry::from(p))
-        .filter(|f| f.is_env_file())
-        .collect();
-
-    // Adds files to paths if they should be included
-    if let Some(included) = args.values_of("include") {
-        included
-            .filter_map(|f| fs::canonicalize(f).ok())
-            .filter_map(|p| FileEntry::from(p))
-            .for_each(|f| paths.push(f));
-    }
-
-    // Removes files from paths if they should be excluded
-    if let Some(excluded) = args.values_of("exclude") {
-        let excluded_paths: Vec<PathBuf> =
-            excluded.filter_map(|f| fs::canonicalize(f).ok()).collect();
-
-        paths.retain(|f| !excluded_paths.contains(&f.path));
-    }
-
-    paths.sort();
-    paths.dedup();
-
-    Ok(paths)
 }